Move server selection to Settings and allows URLs
With this change, the user can select a server before initiating a
connection for the first time (important if the default server
is not working for any reason).  This change also allows advanced
users to use arbitrary DOH servers, not just the two presets.

This change removes the server control from the main screen.
This is good for visual consistency, but if users find it harder
to discover this option then that could be an unfortunate side-effect.

Fixes #21