State machine revamp
This moves connection state tracking to a model-view-controller
arrangement, and exposes additional useful information to the
user, such as whether queries have recently been failing.  This
is a prerequisite for proposed UI changes that would cause all
users to connect to the default server, because some users may
not be able to reach that server, and we currently have no clear
way to indicate the problem.

Fixes #49