Use tun2socks and sockslib to implement Intra
On Android M (API 23) and later, this version uses a full-scope
VPN to intercept all traffic, instead of a split-horizon VPN that
only covers the DNS server.  The VPN re-emits all traffic unmodified
except for DNS, which is routed through DOH.

This includes a subset of the sockslib source code, binaries for tun2socks (taken
unmodified from Outline), and a suite of new classes to implement the
new VPN functionality.

Fixes #51
Fixes #19