Fix a crash in sockslib's StreamPipe
We've seen several crash reports from users related to this issue.
The crash is an IndexOutOfRangeException in StreamPipe.doTransfer.
It occurs because the list of listeners is modified during iteration,
so the iteration runs past the end of the list.

This pattern appears throughout StreamPipe.  This change resolves the
problem by making a shallow copy of pipeListeners before iterating
through the list.

TODO: Upstream into sockslib.