Fix autostart-on-boot and autostart-on-launch
Previously, autostart-on-launch would never call VpnService.prepare,
and autostart-on-boot would fail if VPN permission had been lost.
The former is a problem when the user launches Intra before autostart
completes, and the latter is a problem on older devices that reset
the VPN permission on reboot.

This change fixes autostart-on-launch, and makes autostart-on-boot
launch the app if VPN permission has been lost, so that the user
can re-permission Intra.

Fixes #105