Add caching
This change adds DoH-compliant caching and TTL rewriting.
It uses Caffeine, a cache framework that supports custom
expiration times for each entry and load coalescing, so
that a query can be fulfilled from cache if a load action
is in progress for that query.

Caching is disabled by default, subject to a Firebase
remote configuration experiment.

NOTE: Currently this change is blocked by a compatibility
bug in Caffeine: https://github.com/ben-manes/caffeine/issues/293

Fixes #90