Enable proguard for release builds
This is the recommended build configuration for Android.
It's also required because otherwise (thanks to Guava) we
have more than 65536 methods, which is the limit for one
Android DEX file, and support for multi-DEX APKs on older
Android versions is limited.  Proguard removes all the
unused methods, keeping us under the DEX file limit (and
also reducing the APK size).