Switch Google to RFC DoH and remove special cases
With this change, Intra only speaks standard RFC 8484
DoH, and Google Public DNS is a DoH server like any
other.

This simplifies the DNS lookup behavior considerably,
and allows us to consider alternative DoH client
implementations.

The ServerConnection interface is now redundant, and
will be removed in a subsequent change.