Avoid silent failures when Go-DOH setup fails
Go-DOH setup currently throws an exception if `url` is null,
which is the default value on a fresh install.  If Go-DOH setup
throws an exception, the VPN silently fails to start and the error is
logged and discarded, leaving the UI in an ambiguous state.

This was causing total failure when using Go-DOH in fresh installs
and installs that had never adjusted the server setting.