fix(crash): mark the VPN stop notification as immutable
Android S+ requires all PendingIntents to be marked as mutable
or immutable (usually the latter).  This change adds this
annotation to the PendingIntent used to surface Intra when it
is disabled.  This occurs most commonly when another VPN is
started.  Without this fix, Intra crashes when another VPN
is started during a session.