Refactoring dbip_shrink.py
In python It's more recommended that open files with a context manager so in cases of any unexpected error, context manager can handle situation and close the files properly. it is also close the file or files automatically after completed operation. and also cleaner and more readable code were provided.