fix: cancel all outgoing requests when disconnect to prevent ANR (#501)
In this PR, I fixed an ANR issue by adding a `context.Context` object to the `doh.Transport.Query` method. This `ctx` will be passed to all network related calls, and it will be cancelled when user `Disconnect`s.

But if we just simply add `ctx` to the method, `gomobile` requires we export all related packages (e.g., `time`) that are used by `context.Context`, which is not applicable. Therefore we refactored the go code structure and introduced a new `backend` package that will be the only interface that Java code can use.

In addition, I retired `github.com/eycorsican/go-tun2socks/common/log` and introduced our own `logging` package.